Update test cases for conversation list
